+++ packages/react/.storybook/preview.js
@@ -20,7 +20,11 @@ addParameters({
     theme: theme,
     showRoots: true,
     storySort: (a, b) =>
-      a[1].kind === b[1].kind ? 0 : a[1].id.localeCompare(b[1].id, undefined, { numeric: true }),
+      a[1].kind.replace(/☢️-|🚫-|⚠️-/i, '') === b[1].kind.replace(/☢️-|🚫-|⚠️-/i, '')
+        ? 0
+        : a[1].id
+            .replace(/☢️-|🚫-|⚠️-/i, '')
+            .localeCompare(b[1].id.replace(/☢️-|🚫-|⚠️-/i, ''), undefined, { numeric: true }),
   },
 });