+++ packages/react/src/components/Dashboard/DashboardHeader.story.jsx
@@ -6,10 +6,17 @@ import { TrashCan20 } from '@carbon/icons-react';
 import { action } from '@storybook/addon-actions';
 import { DatePicker, DatePickerInput } from 'carbon-components-react';
 
+import StoryNotice, { deprecatedStoryTitle } from '../../internal/StoryNotice';
+
 import DashboardHeader from './DashboardHeader';
 
 export default {
-  title: 'Watson IoT/Dashboard Header (Deprecated)',
+  title: 'Watson IoT/🚫 Dashboard Header',
+};
+
+export const Deprecated = () => <StoryNotice componentName="DashboardHeader" />;
+Deprecated.story = {
+  name: deprecatedStoryTitle,
 };
 
 export const Basic = () => {